+++ examples/foundational/26h-gemini-multimodal-live-vertex.py
@@ -17,8 +17,8 @@ from pipecat.pipeline.runner import PipelineRunner
 from pipecat.pipeline.task import PipelineParams, PipelineTask
 from pipecat.runner.types import RunnerArguments
 from pipecat.runner.utils import create_transport
-from pipecat.services.gemini_live.gemini import GeminiLiveLLMService
-from pipecat.services.gemini_live.vertex import GeminiLiveVertexLLMService
+from pipecat.services.google.gemini_live.gemini import GeminiLiveLLMService
+from pipecat.services.google.gemini_live.vertex import GeminiLiveVertexLLMService
 from pipecat.transports.base_transport import BaseTransport, TransportParams
 from pipecat.transports.daily.transport import DailyParams
 from pipecat.transports.websocket.fastapi import FastAPIWebsocketParams
